// generated/src/app/api-models/member-master-ud-arc.model.ts
/* Copyright (c) 2021 . All Rights Reserved. */

export class MemberMasterUdArc {
  residencyLocationCode: string;
  comments: string;
  dubaiWorkLocationCode: string;
  uaeVisaStartDate: Date;
  updateProcess: string;
  updateUser: string;
  updateDatetime: Date;
  insertProcess: string;
  insertUser: string;
  insertDatetime: Date;
  userDate10: Date;
  userDate9: Date;
  userDate8: Date;
  userDate7: Date;
  userDate6: Date;
  userDate5: Date;
  userDate4: Date;
  userDate3: Date;
  userDate2: Date;
  userDate1: Date;
  userDefined25: string;
  userDefined24: string;
  userDefined23: string;
  userDefined22: string;
  userDefined21: string;
  userDefined20: string;
  userDefined19: string;
  userDefined18: string;
  userDefined17: string;
  userDefined16: string;
  userDefined15: string;
  userDefined14: string;
  userDefined13: string;
  userDefined12: string;
  userDefined11: string;
  userDefined10: string;
  userDefined9: string;
  userDefined8: string;
  userDefined7: string;
  userDefined6: string;
  userDefined5: string;
  userDefined4: string;
  userDefined3: string;
  userDefined2: string;
  userDefined1: string;
  preferredCurrency: string;
  ksaProfession: string;
  sponsorIdType: string;
  sponsorIdIssueLoc: string;
  sponsorId: string;
  governmentIdType: string;
  governmentIdIssueLoc: string;
  governmentIdExpireDate: Date;
  governmentId: string;
  emirateIdIssueLoc: string;
  emirateIdExpireDate: Date;
  emirateIdNo: string;
  ksaNationality: string;
  passportIssueLoc: string;
  passportExpireDate: Date;
  passportNumber: string;
  personNumber: string;
  subscriberId: string;
  seqSubsId: number;
  seqMembId: number;
  changeType: string;
  changeDateTime: Date;
}